[HUDI-1253] Merge two TestHoodieSnapshotExporter class in hudi-utilities
